TrackListLocal добавлена в граф навигации мелкие изменения в ui решена проблема с application context небольшой рефакторинг
