+++ app/src/main/java/com/vikvita/music_player/App.kt
@@ -4,15 +4,28 @@ import android.app.Application
 import com.vikvita.music_player.data.di.DataGraph
 import com.vikvita.music_player.di.AppBindingModule
 import com.vikvita.music_player.di.AppProvidingModule
+import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
 
 @Singleton
 @Component(modules = [DataGraph::class,AppBindingModule::class,AppProvidingModule::class])
 interface ApplicationComponent {
+    @Component.Builder
+    interface Builder {
+        fun build(): ApplicationComponent
+
+        @BindsInstance
+        fun application(application: Application): Builder
+    }
     fun inject(mainActivity: MainActivity)
 }
 
 class App:Application() {
-    val appComponent = DaggerApplicationComponent.create()
+   lateinit var appComponent:ApplicationComponent
+
+    override fun onCreate() {
+        super.onCreate()
+        appComponent=DaggerApplicationComponent.builder().application(this).build()
+    }
 }
